again refactor component Register
