+++ src/Ui/Input/UserFullNameInput/UserNameInput/NameInput.module.scss
@@ -1,0 +1,49 @@
+@import "/src/Styles/Variables.module.scss";
+
+.field {
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+
+  gap: 8px;
+  margin-bottom: 24px;
+
+  &__label {
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 22.4px;
+
+    color: $color-darkTheme-003;
+
+    &--invalid {
+      color: #e32227;
+    }
+  }
+
+  &__input {
+    width: 100%;
+    height: 40px;
+
+    padding-top: 12px;
+    padding-right: 16px;
+    padding-bottom: 12px;
+    padding-left: 16px;
+
+    border: 1px solid $color-darkTheme-002;
+    border-radius: 8px;
+
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 16px;
+
+    color: $color-lightTheme-001;
+
+    &::placeholder {
+      font-weight: 400;
+      font-size: 16px;
+      line-height: 16px;
+
+      color: $color-darkTheme-005;
+    }
+  }
+}
