+++ src/Ui/Input/UserPasswordInput/ShowPassword/UserShowRepeatPassword/UserShowRepeatPassword.module.scss
@@ -1,0 +1,16 @@
+@import "/src/Styles/Variables.module.scss";
+
+.visibilityPasswordContainer {
+  display: flex;
+  align-items: center;
+
+  gap: 12px;
+
+  &__label {
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 22.4px;
+
+    color: $color-darkTheme-003;
+  }
+}
